feature: api chatbot
se implemento el backend del chatbot con OpenRouter
