+++ .gitignore
@@ -35,3 +35,7 @@ out/
 
 ### VS Code ###
 .vscode/
+
+# Archivos de configuración sensibles
+.env
+src/main/resources/application.properties
