+++ src/main/java/com/keepu/webAPI/service/OpenRouterService.java
@@ -1,0 +1,40 @@
+package com.keepu.webAPI.service;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
+
+import java.util.Map;
+
+@Service
+public class OpenRouterService {
+
+    @Value("${openrouter.api.key}")
+    private String apiKey;
+
+    private final RestTemplate restTemplate = new RestTemplate();
+
+    public String infer(String prompt) {
+        String url = "https://openrouter.ai/api/v1/chat/completions";
+        HttpHeaders headers = new HttpHeaders();
+        headers.setContentType(MediaType.APPLICATION_JSON);
+        headers.set("Authorization", "Bearer " + apiKey);
+
+        Map<String, Object> body = Map.of(
+                "model", "mistralai/mistral-7b-instruct",
+                "messages", new Object[] {
+                        Map.of("role", "user", "content",
+                                "Solo responde preguntas de finanzas y responde en máximo 200 caracteres. " + prompt)
+                },
+                "max_tokens", 200
+        );
+
+        HttpEntity<Map<String, Object>> request = new HttpEntity<>(body, headers);
+        ResponseEntity<String> response = restTemplate.postForEntity(url, request, String.class);
+        return response.getBody();
+    }
+}
