+++ src/main/resources/application.properties
@@ -29,8 +29,7 @@ app.jwt.token.expiration=28800000
 cashflow.openapi.dev-url=http://localhost:8080
 cashflow.openapi.prod-url=http://cashflow-api.com
 
-huggingface.api.key=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
-
+openrouter.api.key=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
 # Para ver los logs de debug de tu aplicación
 logging.level.tu.paquete.security=DEBUG
 logging.level.org.springframework.security=DEBUG
